+++ CHANGELOG.md
@@ -30,7 +30,8 @@
     * The old functionality is still available with the mongodb-legacy library
 * Breaking: Removed Junit 5 Vintage engine and Junit 4 (can be added in projects that need it)
 * Fixed vulnerabilities: CVE-2023-52428(nimbus-jose-jwt), CVE-2024-29857,CVE-2024-30171,CVE-2024-30172,CVE-2024-34447 (bouncycastle), CVE-2024-28752 (Apache CXF)
-    
+* Added support for multiple demo data sql scripts, use UTF-8 as default encoding
+
 # 1.35
 * Update dependencies
     * Spring 5.3.31